requirements.txt
google-auth>=2.22.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
pandas>=2.1.0
openpyxl>=3.1.2
PyYAML>=6.0.1
httpx>=0.25.0
streamlit>=1.30.0
